Moved SECOND, HOUR and DAY to pywws.constants
