+++ src/pywws/__init__.py
@@ -1,3 +1,3 @@
-__version__ = '15.03.0.dev1280'
-_release = '1280'
-_commit = '9d77f55'
+__version__ = '15.03.0.dev1281'
+_release = '1281'
+_commit = 'c12fc80'
